Refactor and align CLI
